refactor: 유저세팅, 오늘의 강의 api를 한번 호출하도록 개선
