+++ src/components/App.jsx
@@ -11,24 +11,31 @@ import MyPage from "./MyPage";
 import Nav from "../Nav";
 import PasswordError from "../error/PasswordError";
 import React from "react";
+import { TodayLectureProvider } from "../contexts/TodayLectureContext";
+import { UserSettingsProvider } from "../contexts/UserSettingsContext";
 
 const App = () => {
   return (
-    <div>
-      <Nav />
-      <Routes>
-        <Route path="/home" element={<Home />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/passwordError" element={<PasswordError />} />
-        <Route path="/courses/:courseCode" element={<Courses />} />
-        <Route path="/courses/:courseCode/notices" element={<CourseNotice />} />
-        <Route path="/courses" element={<CoursesList />} />
-        <Route path="/mypage" element={<MyPage />} />
-        <Route path="/about" element={<About />} />
-        <Route path="/LoginError" element={<AuthErrorPage />} />
-        <Route path="*" element={<Navigate to="/not-found" replace />} />
-      </Routes>
-    </div>
+    <UserSettingsProvider>
+      <TodayLectureProvider>
+        <Nav />
+        <Routes>
+          <Route path="/home" element={<Home />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="/passwordError" element={<PasswordError />} />
+          <Route path="/courses/:courseCode" element={<Courses />} />
+          <Route
+            path="/courses/:courseCode/notices"
+            element={<CourseNotice />}
+          />
+          <Route path="/courses" element={<CoursesList />} />
+          <Route path="/mypage" element={<MyPage />} />
+          <Route path="/about" element={<About />} />
+          <Route path="/LoginError" element={<AuthErrorPage />} />
+          <Route path="*" element={<Navigate to="/not-found" replace />} />
+        </Routes>
+      </TodayLectureProvider>
+    </UserSettingsProvider>
   );
 };
